// data/github/current/ui/packages/copilot-for-business/types.ts
export const CopilotForBusinessSeatPolicy = {
  Disabled: 'disabled',
  EnabledForAll: 'enabled_for_all',
  EnabledForSelected: 'enabled_for_selected',
} as const

export type CopilotForBusinessSeatPolicy =
  (typeof CopilotForBusinessSeatPolicy)[keyof typeof CopilotForBusinessSeatPolicy]

export const SeatType = {
  OrganizationInvitation: 'OrganizationInvitation',
  Team: 'Team',
  User: 'User',
  Organization: 'Organization',
  All: 'All',
  EnterpriseTeam: 'EnterpriseTeam',
} as const

export type SeatType = (typeof SeatType)[keyof typeof SeatType]

export type SeatBreakdown = {
  seats_assigned: number
  seats_billed: number
  seats_pending: number
  description: string
}

export type CopilotForBusinessTrial = {
  started: boolean
  ended: boolean
  has_trial: boolean
  upgradable: boolean
  cancelable: boolean
  days_left: number
  started_at: string
  ends_at: string
  trial_length: number
  active: boolean
  expired: boolean
  pending: boolean
  copilot_plan: string
  belongs_to_trial_business_account: boolean
}

export type CopilotForBusinessPayload = {
  policy: CopilotForBusinessSeatPolicy
  public_code_suggestions_configured: boolean
  seats: {
    seats: CopilotSeatAssignment[]
    count: number
    pending_requests?: {
      requesters: Requester[]
      count: number
    }
    licenses: CopilotLicenseIdentifiers
  }
  organization: CopilotForBusinessOrganization
  business?: CopilotForBusinessBusiness
  seat_breakdown: SeatBreakdown
  seat_assignments: number
  business_trial?: CopilotForBusinessTrial
  render_trial_expired_banner: boolean
  render_copilot_insights_banner: boolean
  can_add_teams: boolean
  members_count: number
  can_allow_to_assign_seats_on_business: boolean
  next_billing_date: string
  plan_text: 'Business' | 'Enterprise'
  featureRequestInfo: FeatureRequestInfo
  render_pending_downgrade_banner: boolean
  display_activity: boolean
  adoption_metrics: {
    total: number
    active: number
    inactive: number
    dormant: number
  }
}

export type CopilotLicenseIdentifiers = {
  user_ids: number[]
  team_ids: number[]
  invite_user_ids: number[]
  invite_emails: string[]
}

export type CopilotForBusinessOrganization = {
  name: string
  id: number
  billable: boolean
  has_seat: boolean
  add_seat_link: string | null
  copilot_enabled: boolean
}

export type CopilotForBusinessBusiness = {
  name: string
  slug: string
}

export type CopilotForBusinessTeam = {
  id?: number
  login: string
  slug: string
  avatar_url: string
  combined_slug: string
  member_count: number
  member_ids: number[]
}

export type CopilotEnterpriseTeam = {
  id: number
  login: string
  slug: string
  mapping_id?: number // this points to the external group mapping id
  member_count: number
  member_ids: number[]
}

export type CopilotSeatAssignable = {
  id: number
  login?: string
  avatar_url?: string
  display_name?: string
  slug?: string
  combined_slug?: string
  member_count?: number
  member_ids?: number[]
  email?: string
  invitee?: CopilotSeatAssignable
  invitee_id?: number
  mapping_id?: number
}

export type CopilotSeatAssignment<
  Type extends SeatType = SeatType,
  Assignable extends CopilotSeatAssignable = CopilotSeatAssignable,
> = {
  assignable: Assignable
  assignable_type: Type
  id?: number | null
  invitation_date?: string
  invitation_expired?: boolean
  last_activity_at?: string
  pending_cancellation_date?: string | null
  access_revoked_at?: string | null
}
// ...
export type FeatureRequestInfo = {
  showFeatureRequest: boolean
  alreadyRequested: boolean
  dismissed: boolean
  featureName: string
  requestPath: string
  isEnterpriseRequest?: boolean
  dismissedAt?: string
  billingEntityId?: string
  latestUsernameRequests: string[]
  amountOfUserRequests: number
}

export type Requester = {
  id: number
  display_login: string
  profile_name: string | null
  requested_at?: string
}